backend/requirements.txt
supabase
google-generativeai
pypdf
python-dotenv